Rip out GET query on backend
